% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iot1clickprojects_operations.R
\name{iot1clickprojects_list_tags_for_resource}
\alias{iot1clickprojects_list_tags_for_resource}
\title{Lists the tags (metadata key/value pairs) which you have assigned to the
resource}
\usage{
iot1clickprojects_list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The ARN of the resource whose tags you want to list.}
}
\description{
Lists the tags (metadata key/value pairs) which you have assigned to the
resource.
}
\section{Request syntax}{
\preformatted{iot1clickprojects$list_tags_for_resource(
  resourceArn = "string"
)
}
}

\keyword{internal}
